refactor: move alert and info to svgr
