--- packages/interface/src/components/icons/Info.tsx
@@ -1,26 +1,0 @@
-import React from 'react';
-
-interface InfoProps {
-	/**
-	 * Add additional classes to the SVG element
-	 */
-	className?: string;
-}
-
-export function Info(props: InfoProps) {
-	return (
-		<svg
-			className={props.className}
-			width="32"
-			height="32"
-			viewBox="0 0 32 32"
-			fill="none"
-			xmlns="http://www.w3.org/2000/svg"
-		>
-			<path
-				d="M16 3C13.4288 3 10.9154 3.76244 8.77759 5.1909C6.63975 6.61935 4.97351 8.64968 3.98957 11.0251C3.00563 13.4006 2.74819 16.0144 3.2498 18.5362C3.75141 21.0579 4.98953 23.3743 6.80762 25.1924C8.6257 27.0105 10.9421 28.2486 13.4638 28.7502C15.9856 29.2518 18.5995 28.9944 20.9749 28.0104C23.3503 27.0265 25.3807 25.3603 26.8091 23.2224C28.2376 21.0846 29 18.5712 29 16C28.9934 12.5542 27.6216 9.25145 25.1851 6.81491C22.7486 4.37837 19.4458 3.00661 16 3ZM15.75 9C16.0467 9 16.3367 9.08797 16.5834 9.2528C16.83 9.41762 17.0223 9.65189 17.1358 9.92597C17.2494 10.2001 17.2791 10.5017 17.2212 10.7926C17.1633 11.0836 17.0204 11.3509 16.8107 11.5607C16.6009 11.7704 16.3336 11.9133 16.0426 11.9712C15.7517 12.0291 15.4501 11.9994 15.176 11.8858C14.9019 11.7723 14.6676 11.58 14.5028 11.3334C14.338 11.0867 14.25 10.7967 14.25 10.5C14.25 10.1022 14.408 9.72064 14.6893 9.43934C14.9706 9.15804 15.3522 9 15.75 9ZM17 23H16C15.7348 23 15.4804 22.8946 15.2929 22.7071C15.1054 22.5196 15 22.2652 15 22V16C14.7348 16 14.4804 15.8946 14.2929 15.7071C14.1054 15.5196 14 15.2652 14 15C14 14.7348 14.1054 14.4804 14.2929 14.2929C14.4804 14.1054 14.7348 14 15 14H16C16.2652 14 16.5196 14.1054 16.7071 14.2929C16.8946 14.4804 17 14.7348 17 15V21C17.2652 21 17.5196 21.1054 17.7071 21.2929C17.8946 21.4804 18 21.7348 18 22C18 22.2652 17.8946 22.5196 17.7071 22.7071C17.5196 22.8946 17.2652 23 17 23Z"
-				fill="fill-current"
-			/>
-		</svg>
-	);
-}
